requirements.txt
matplotlib>=3.10.3
numpy>=2.0.0      
ollama>=0.6.1
openai>=1.0.0
python-dotenv>=1.2.1
requests>=2.32.3
shapely>=2.1.1
tiktoken>=0.9.0